modernizing c types part H

+++ libs/gretchen/src/helper/src/hashmap.c
@@ -336,7 +336,7 @@ int8_t hashmap_iterate(map_t in, PFany f, any_t item) {
 	for(i = 0; i< m->table_size; i++)
 		if(m->data[i].in_use != 0) {
 			any_t data = (any_t) (m->data[i].data);
-			int status = f(item, data);
+			uint8_t status = f(item, data);
 			if (status != MAP_OK) {
 				return status;
 			}
